do not overwrite outputfiles

+++ twitter_monitor/basic_stream.py
@@ -137,6 +137,9 @@ def get_tweepy_auth(twitter_api_key,
 def construct_listener(outfile=None):
     """Create the listener that prints tweets"""
     if outfile is not None:
+        if os.path.exists(outfile):
+            raise IOError("File %s already exists" % outfile)
+        
         outfile = open(outfile, 'wb')
 
     return PrintingListener(out=outfile)
